Update post service

+++ src/domain/useCases/post/deletePost.ts
@@ -1,9 +1,17 @@
 import PostService from "src/app/services/postService";
+import { BadRequestError } from "src/presentation/utils/customError";
 
 export default class DeletePost {
   constructor(private postService: PostService) {}
 
   async execute(id: string): Promise<void> {
+
+    const post = await this.postService.get(id);
+
+    if (!post) {
+      throw new BadRequestError(`Post ${id} not found`);
+    }
+
     return await this.postService.delete(id);
   }
 }
